% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FenotipoCategorias.R
\name{FenotipoCategorias}
\alias{FenotipoCategorias}
\title{Carga archivo de resultados fenotipicos (AST)
y selecciona columnas por Categoria RAM}
\usage{
FenotipoCategorias(archivo_fenotipo)
}
\arguments{
\item{archivo_fenotipo}{Ruta del archivo de entrada de datos fenotipicos (AST)}
}
\value{
Data Frame con los IDs y la columna de resultados de AST del antimicrobiano seleccionado
}
\description{
Carga el archivo de ASTs.
Proporciona las hojas existentes en el archivo de excel al ususario.
Proporciona una lista de antimicrobianos disponibles y pide al ususario sellecionar un antimicrobiano.
Selecciona la columna de identificadores de aislados (columna ID).
Selecciona la columna de resultados por AST del antimicrobiano seleccionado.
}
