% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotipoCategorias.R
\name{GenotipoCategorias}
\alias{GenotipoCategorias}
\title{Carga archivo de resultados genotipicos de WGS (por ResFinder) y
obtiene perfiles de genes RAM}
\usage{
GenotipoCategorias(archivo_genotipo)
}
\arguments{
\item{archivo_genotipoRuta}{del archivo de genes RAM (por ResFinder)}
}
\value{
Una tabla de conteos de resultados de AST correspondientes con perfiles de genes RAM y sus IDs
}
\description{
Carga el archivo de genes RAM
Limpia el archivo de genes RAM
Filtra el archivo de genes RAM
Obtiene Especificidad
Obtiene Sensibilidad
Obtiene conteo de perfiles de genes RAM (combinaciones de genes por aislados)
Relaciona y compara resultados de genes RAM con ASTs
Obtiene tabla final de conteos de AST con ciertos perfiles de genes RAM (o ausencia de estos) y sus IDs
}
