% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuardarCategoriasRAM.R
\name{GuardarCategoriasRAM}
\alias{GuardarCategoriasRAM}
\title{Guarda Perfiles RAM y Tabla Final por categorias
(conteos e IDs de perfiles RAM y ASTs)}
\usage{
GuardarCategoriasRAM(ruta, nombre)
}
\arguments{
\item{ruta}{Ruta del directorio donde se creara el directorio para guardar resultados}
}
\value{
Crea directorio de categoria de antimicrobiano correspondiente y
guarda dentro archivo de perfiles y tabla final.
}
\description{
Guarda los resultados mas importantes que son:
Archivo Perfiles que contiene perfiles de Genes RAM y sus conteos (asi como su ausencia)
Archivo Final que contiene la tabla de resultados en conteos de aislados con cierto perfil
de genes RAM y sus ausencia, relacionados a resultados fenotipicos.
}
